requirements.txt
pandas
transformers
numpy
docx2txt
python-dotenv
einops